requirements.txt
opencv-python>=4.8.0
ultralytics>=8.0.0
torch>=2.0.0
Pillow>=10.0.0
openvino>=2024.0.0
